+++ quam_builder/architecture/quantum_dots/examples/quam_qd_example.py
@@ -183,6 +183,7 @@ if qdac_connect:
                 "channel_method": "channel",
                 "accessor": "dc_constant_V",
                 "is_qdac": True,
+                "close_method": "close",
             }
         }
     )